Add files via upload
A library management system software with capabilities of barcoding and RFID helps in scanning the barcode while lending or returning books. Management of the catalogue and inventory by the system makes the process accountable. The feature of the software assists in inventory and circulation management of the books. Books in stock, on the shelves, in circulation, missing or to be ordered can be tracked by the system. Acquiring new assets becomes easier with the software.